backend/pyproject.toml
[project]
name = "jetlag-backend"
version = "0.1.0"
description = "Jet Lag: The Game — real-time companion backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.34.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.7.0",
  "email-validator>=2.2.0",
  "structlog>=25.1.0",
  "httpx>=0.28.0",
  "redis[hiredis]>=5.2.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "aiosqlite>=0.20.0",
  "alembic>=1.14.0",
  "pyjwt[crypto]>=2.10.0",
  "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
  "pytest-cov>=6.0.0",
  "ruff==0.15.10",
  "ssort>=0.14.0",
  "ty==0.0.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jetlag"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["jetlag"]

[tool.coverage.report]
fail_under = 70
show_missing = true
